+++ crates/base/core/src/data/layout/buffer/linear.rs
@@ -1,27 +1,33 @@
 // devela_base_core::data::layout::buffer::linear
 //
 //! Defines [`buffer_linear!`].
+//!
+//! > A semantic machine that overlays occupancy semantics over contiguous storage.
 //
 
 #[doc = crate::_tags!(construction data_structure)]
 /// Defines a linear buffer type over contiguous storage.
 #[doc = crate::_doc_location!("data/layout")]
 ///
-/// The generated type represents a growable logical prefix over a
+/// The generated type represents a growable logical range over a
 /// contiguous storage backend, using an index type to track length
 /// and element positions.
 ///
-/// ## Ownership modes
+/// ## Ownership regimes
 ///
-/// The macro supports two ownership modes:
+/// The macro supports three regimes:
 ///
-/// - **Owned** (default)
-///   The buffer owns its storage backend.
-///   The optional `owned` specifier can be omitted.
+/// - **static** (default)
+///   The buffer owns its static storage backend.
+///   The optional `static` specifier can be omitted.
 ///
 /// - **View**
 ///   The buffer is a non-owning view over externally provided storage.
 ///   The `view` specifier must be made explicit.
+///
+/// - **Alloc**
+///   The buffer owns its allocated storage backend.
+///   The `alloc` specifier must be made explicit.
 ///
 /// ## Index type requirements
 ///
@@ -38,7 +44,7 @@
 ///
 /// Backends are opt-in and selected by listing them after the struct declaration.
 ///
-/// ### Owned backends
+/// ### Static backends
 ///
 /// - **`array`**
 ///   Fully initialized fixed-size array (`[T; CAP]`).
@@ -60,19 +66,24 @@
 ///
 /// - **`slice`**
 ///   Shared slice (`&[T]`).
-///   Read-only view over a contiguous prefix.
+///   Read-only view over a contiguous logical range.
+///
+/// ### Alloc backends
+///
+/// - **`vec`**
+///   Continuous growable `Vec<T>` array.
 ///
 /// Only the implementations explicitly enabled are generated.
 /// Constructors and methods depend on the selected backends.
 ///
 /// ## Examples
 ///
-/// ### Owned buffer (default mode)
+/// ### Static buffer (default)
 /// ```
 /// # #![cfg_attr(nightly_doc, feature(doc_cfg))] // reason = _devela_policy! emmiting doc(cfg)
 /// # use devela_base_core::buffer_linear;
 /// buffer_linear!(
-///     /// Owned linear buffer.
+///     /// Static linear buffer.
 ///     pub struct BufferU8: (u8); array
 /// );
 /// let mut buf = BufferU8::<i32, [i32; 8]>::new_init();
@@ -101,19 +112,22 @@
 /// buffer_linear!(pub struct BufferSplit: view (u8););
 /// buffer_linear!(impl BufferSplit: view (u8); slice, slice_mut);
 /// ```
-/// See also: [`BufferExample`][crate::BufferExample],
-/// [`BufferViewExample`][crate::BufferViewExample].
+/// See also: [`BufferStaticExample`][crate::data::layout::BufferStaticExample],
+/// [`BufferViewExample`][crate::data::layout::BufferViewExample],
+/// [`BufferAllocExample`].
+#[doc = crate::doclink!(custom devela
+    "[`BufferAllocExample`]" "data/layout/struct.BufferAllocExample.html")]
 //
 // NOTE: The index type is passed as a token group to allow complex or path-qualified types.
 #[macro_export]
 #[cfg_attr(cargo_primary_package, doc(hidden))]
 macro_rules! buffer_linear {
     (
-    // OWNED (array, uninit, option)
+    // STATIC (array, uninit, option)
     // struct definition + optional implementations
 
         $(#[$struct_attr:meta])*                                 // attributes
-        $vis:vis struct $name:ident : $(owned)? ($($I:tt)+);     // visibility, name, index type
+        $vis:vis struct $name:ident : $(static)? ($($I:tt)+);    // visibility, name, index type
         $($rest:tt)*                                             // impls
     ) => {
         $(#[$struct_attr])*
@@ -123,14 +137,14 @@ macro_rules! buffer_linear {
             len: $crate::MaybeNiche<$($I)+>,
             _m: $crate::PhantomData<T>,
         }
-        $crate::buffer_linear!(%impl_common_owned $name, $($I)+, $crate::niche_prim![$($I)+]);
-        $crate::buffer_linear!(%impls_owned
+        $crate::buffer_linear!(%impl_common_static $name, $($I)+, $crate::niche_prim![$($I)+]);
+        $crate::buffer_linear!(%impls_static
             $name : $($I)+, $crate::niche_prim![$($I)+] ; $($rest)*);
     };
     ( // implementations only
-        impl $name:ident : $(owned)? ($($I:tt)+) ; $($rest:tt)*  // for name, index type, impls
+        impl $name:ident : $(static)? ($($I:tt)+) ; $($rest:tt)*  // for name, index type, impls
     ) => {
-        $crate::buffer_linear!(%impls_owned
+        $crate::buffer_linear!(%impls_static
             $name : $($I)+, $crate::niche_prim![$($I)+] ; $($rest)*);
     };
     (
@@ -158,34 +172,58 @@ macro_rules! buffer_linear {
         $crate::buffer_linear!(%impls_view
             $name : $($I)+, $crate::niche_prim![$($I)+] ; $($rest)*);
     };
+    (
+    // ALLOC (vec)
+    // struct definition + optional implementations
+
+        $(#[$struct_attr:meta])*                                 // attributes
+        $vis:vis struct $name:ident : alloc ($($I:tt)+);         // visibility, name, index type
+        $($rest:tt)*                                             // impls
+    ) => {
+        $(#[$struct_attr])*
+        #[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
+        $vis struct $name<T, S> {
+            storage: S,
+            _m: $crate::PhantomData<T>,
+        }
+        $crate::buffer_linear!(%impl_common_alloc $name, $($I)+, $crate::niche_prim![$($I)+]);
+        $crate::buffer_linear!(%impls_alloc
+            $name : $($I)+, $crate::niche_prim![$($I)+] ; $($rest)*);
+    };
+    ( // implementations only
+        impl $name:ident : alloc ($($I:tt)+) ; $($rest:tt)*      // for name, index type, impls
+    ) => {
+        $crate::buffer_linear!(%impls_alloc
+            $name : $($I)+, $crate::niche_prim![$($I)+] ; $($rest)*);
+    };
+
     // --------------------------------------------------------------------------------------------
-    //% owned dispatch
+    //% static dispatch
     (
 
     /* internals */
-     %impls_owned $name:ident : $I:ty, $P:ty ;) => {}; // no impls
-    (%impls_owned $name:ident : $I:ty, $P:ty ; $(#[$i:meta])* $impl:ident) => { // last impl
-        $crate::buffer_linear!(%impl1_owned $name : $I, $P ; $impl);
+     %impls_static $name:ident : $I:ty, $P:ty ;) => {}; // no impls
+    (%impls_static $name:ident : $I:ty, $P:ty ; $(#[$i:meta])* $impl:ident) => { // last impl
+        $crate::buffer_linear!(%impl1_static $name : $I, $P ; $impl);
     };
-    // owned: array
-    (%impls_owned $name:ident : $I:ty, $P:ty ; $(#[$i:meta])* array , $($rest:tt)*) => {
+    // static: array
+    (%impls_static $name:ident : $I:ty, $P:ty ; $(#[$i:meta])* array , $($rest:tt)*) => {
         $crate::buffer_linear!(%impl_array $(#[$i])* $name, $I, $P);
-        $crate::buffer_linear!(%impls_owned $name : $I, $P ; $($rest)*); };
-    (%impl1_owned $(#[$i:meta])* $name:ident : $I:ty, $P:ty; array) => {
+        $crate::buffer_linear!(%impls_static $name : $I, $P ; $($rest)*); };
+    (%impl1_static $(#[$i:meta])* $name:ident : $I:ty, $P:ty; array) => {
         $crate::buffer_linear!(%impl_array $(#[$i])* $name, $I, $P); };
-    // owned: uninit
-    (%impls_owned $name:ident : $I:ty, $P:ty ; $(#[$i:meta])* uninit , $($rest:tt)*) => {
+    // static: uninit
+    (%impls_static $name:ident : $I:ty, $P:ty ; $(#[$i:meta])* uninit , $($rest:tt)*) => {
         $crate::buffer_linear!(%impl_uninit $(#[$i])* $name, $I, $P);
-        $crate::buffer_linear!(%impls_owned $name : $I, $P ; $($rest)*); };
-    (%impl1_owned $(#[$i:meta])* $name:ident : $I:ty, $P:ty; uninit) => {
+        $crate::buffer_linear!(%impls_static $name : $I, $P ; $($rest)*); };
+    (%impl1_static $(#[$i:meta])* $name:ident : $I:ty, $P:ty; uninit) => {
         $crate::buffer_linear!(%impl_uninit $(#[$i])* $name, $I, $P); };
-    // owned: option
-    (%impls_owned $name:ident : $I:ty, $P:ty ; $(#[$i:meta])* option , $($rest:tt)*) => {
+    // static: option
+    (%impls_static $name:ident : $I:ty, $P:ty ; $(#[$i:meta])* option , $($rest:tt)*) => {
         $crate::buffer_linear!(%impl_option $(#[$i])* $name, $I, $P);
-        $crate::buffer_linear!(%impls_owned $name : $I, $P ; $($rest)*); };
-    (%impl1_owned $(#[$i:meta])* $name:ident : $I:ty, $P:ty; option) => {
+        $crate::buffer_linear!(%impls_static $name : $I, $P ; $($rest)*); };
+    (%impl1_static $(#[$i:meta])* $name:ident : $I:ty, $P:ty; option) => {
         $crate::buffer_linear!(%impl_option $(#[$i])* $name, $I, $P); };
-
     //% view dispatch
     (%impls_view $name:ident : $I:ty, $P:ty ;) => {}; // no impls
     (%impls_view $name:ident : $I:ty, $P:ty ; $(#[$i:meta])* $impl:ident) => { // last impl
@@ -203,32 +241,51 @@ macro_rules! buffer_linear {
         $crate::buffer_linear!(%impls_view $name : $I, $P ; $($rest)*); };
     (%impl1_view $(#[$i:meta])* $name:ident : $I:ty, $P:ty ; slice) => {
         $crate::buffer_linear!(%impl_slice $(#[$i])* $name, $I, $P); };
+    //% alloc dispatch
+    (%impls_alloc $name:ident : $I:ty, $P:ty ;) => {}; // no impls
+    (%impls_alloc $name:ident : $I:ty, $P:ty ; $(#[$i:meta])* $impl:ident) => { // last impl
+        $crate::buffer_linear!(%impl1_alloc $name : $I, $P ; $impl);
+    };
+    // alloc: vec
+    (%impls_alloc $name:ident : $I:ty, $P:ty ; $(#[$i:meta])* vec , $($rest:tt)*) => {
+        $crate::buffer_linear!(%impl_vec $(#[$i])* $name, $I, $P);
+        $crate::buffer_linear!(%impls_alloc $name : $I, $P ; $($rest)*); };
+    (%impl1_alloc $(#[$i:meta])* $name:ident : $I:ty, $P:ty; vec) => {
+        $crate::buffer_linear!(%impl_vec $(#[$i])* $name, $I, $P); };
 
     // safe-guards
-    (%impls_owned $name:ident : $_I:ty, $_P:ty ; $(#[$_i:meta])* $impl:ident , $($_r:tt)*) => {
-        compile_error!(concat!( "buffer_linear!: unknown impl `", stringify!($impl), "`"));
+    (%impls_static $name:ident : $_I:ty, $_P:ty ; $(#[$_i:meta])* $impl:ident , $($_r:tt)*) => {
+        compile_error!(concat!( "buffer_linear!: unknown static impl `", stringify!($impl), "`"));
     };
     (%impls_view $name:ident : $_I:ty, $_P:ty ; $(#[$_i:meta])* $impl:ident , $($_r:tt)*) => {
-        compile_error!(concat!( "buffer_linear!: unknown impl `", stringify!($impl), "`"));
+        compile_error!(concat!( "buffer_linear!: unknown view impl `", stringify!($impl), "`"));
+    };
+    (%impls_alloc $name:ident : $_I:ty, $_P:ty ; $(#[$_i:meta])* $impl:ident , $($_r:tt)*) => {
+        compile_error!(concat!( "buffer_linear!: unknown alloc impl `", stringify!($impl), "`"));
     };
 
     /* blocks for common private associated items */
 
-    (%impl_common_owned $name:ident, $I:ty, $P:ty) => {
+    (%impl_common_static $name:ident, $I:ty, $P:ty) => {
         /// Common methods.
         impl<T, S> $name<T, S> {
-            $crate::buffer_linear!(%common_private $name, $I, $P);
-            $crate::buffer_linear!(%common_public $name, $I, $P);
+            $crate::buffer_linear!(%common_tracked $name, $I, $P);
         }
     };
     (%impl_common_view $name:ident, $I:ty, $P:ty) => {
         /// Common methods.
         impl<'a, T, S> $name<'a, T, S> {
-            $crate::buffer_linear!(%common_private $name, $I, $P);
-            $crate::buffer_linear!(%common_public $name, $I, $P);
+            $crate::buffer_linear!(%common_tracked $name, $I, $P);
         }
     };
-    (%common_private $name:ident, $I:ty, $P:ty) => {
+    (%impl_common_alloc $name:ident, $I:ty, $P:ty) => {
+        /// Common methods.
+        impl<T, S> $name<T, S> {
+            $crate::buffer_linear!(%common_delegated $name, $I, $P);
+        }
+    };
+    // common items for tracked len (regimes: static, view)
+    (%common_tracked $name:ident, $I:ty, $P:ty) => {
         /// Constructs a buffer from raw components, assuming all invariants hold.
         #[inline(always)]
         const fn _new(storage: S, len: $crate::MaybeNiche<$I>) -> Self {
@@ -327,9 +384,9 @@ macro_rules! buffer_linear {
         const fn _len_dec(&self) -> $crate::MaybeNiche<$I> {
             $crate::unwrap![ok $crate::MaybeNiche::<$I>::try_from_prim(self.len.prim() - 1)]
         }
-    };
-    // common public query methods
-    (%common_public $name:ident, $I:ty, $P:ty) => {
+
+        /* public methods */
+
         /// Returns the number of elements currently stored in the buffer.
         pub const fn len(&self) -> $I { self.len.repr() }
         /// Returns the number of elements currently stored in the buffer.
@@ -337,8 +394,49 @@ macro_rules! buffer_linear {
         /// Returns `true` if the buffer contains no elements.
         pub const fn is_empty(&self) -> bool { self.len.prim() == 0 }
     };
-    // common items for owned variants (array, uninit, option)
-    (%common_owned $name:ident, $I:ty, $P:ty) => {
+    // common items for delegated len (regimes: alloc)
+    (%common_delegated $name:ident, $I:ty, $P:ty) => {
+        /// Constructs a buffer from raw components, assuming all invariants hold.
+        #[inline(always)]
+        const fn _new(storage: S) -> Self {
+            Self { storage, _m: $crate::PhantomData }
+        }
+
+        /* prim */
+
+        /// Returns the given primitive value as an index type.
+        #[inline(always)]
+        const fn _prim_to_idx(from: $P) -> Result<$I, $crate::InvalidValue> {
+            $crate::unwrap![ok_map? $crate::MaybeNiche::<$I>::try_from_prim(from), |v| v.repr()]
+        }
+        /// Returns the given primitive value as an index type,
+        /// converting invalid inputs to the closest valid number.
+        #[inline(always)]
+        const fn _prim_to_idx_lossy(from: $P) -> $I {
+            $crate::MaybeNiche::<$I>::from_prim_lossy(from).repr()
+        }
+
+        /* usize */
+
+        /// Returns the given usize value as a MaybeNiche wrapped saturated index type.
+        #[inline(always)]
+        const fn _usize_to_idx_sat(from: usize) -> $crate::MaybeNiche<$I> {
+            $crate::MaybeNiche::<$I>::from_usize_saturating(from)
+        }
+        /// Returns the given usize value as a MaybeNiche wrapped index type.
+        // It should not panic since we've already checked the invariants.
+        #[inline(always)]
+        const fn _usize_to_idx(from: usize) -> $crate::MaybeNiche<$I> {
+            $crate::unwrap![ok $crate::MaybeNiche::<$I>::try_from_usize(from)]
+        }
+        /// Returns the given index value as a usize.
+        #[inline(always)]
+        const fn _idx_to_usize(from: $I) -> usize {
+            $crate::unwrap![ok $crate::MaybeNiche(from).try_to_usize()]
+        }
+    };
+    // common items for static variants (array, uninit, option)
+    (%common_static $name:ident, $I:ty, $P:ty) => {
         const _CHECK_INVARIANTS: () = {
             assert!(!$crate::MaybeNiche::<$I>::HAS_NEGATIVE,
                 "buffer_linear! index type must be non-negative");
@@ -385,8 +483,10 @@ macro_rules! buffer_linear {
         /// Returns `true` if the buffer has reached its capacity.
         pub const fn is_full(&self) -> bool { Self::_idx_eq(self.len(), self.capacity()) }
     };
-    // common items for array, uninit & and slice_mut variants
-    (%common_array_uninit_slice_mut $name:ident, $I:ty, $P:ty) => {
+    // common items for variants: array, uninit, slice_mut & vec.
+    (%common_iter_visit $name:ident, $I:ty, $P:ty) => {
+        /* iteration */
+
         /// Iterates over the initialized elements.
         pub fn iter(&self) -> impl Iterator<Item = &T> {
             self.as_slice().iter()
@@ -407,19 +507,19 @@ macro_rules! buffer_linear {
             for x in self.as_mut_slice() { f(x); }
         }
 
-        /// Visits the initialized prefix as a shared slice of `Some(T)`,
+        /// Visits the active logical range as a shared slice of `Some(T)`,
         /// without exposing borrow identity.
         pub fn visit_slice<F, R>(&self, f: F)
             -> R where for<'v> F: FnOnce(&'v [T]) -> R { f(self.as_slice())
         }
-        /// Visits the initialized prefix as an exclusive slice of `Some(T)`,
+        /// Visits the active logical range as an exclusive slice of `Some(T)`,
         /// without exposing borrow identity.
         pub fn visit_mut_slice<F, R>(&mut self, f: F)
             -> R where for<'v> F: FnOnce(&'v mut [T]) -> R { f(self.as_mut_slice())
         }
     };
 
-    /* owned variants (array, uninit, option) */
+    /* static variants (array, uninit, option) */
 
     (%impl_array $(#[$impl_attr:meta])* $name:ident, $I:ty, $P:ty) => {
         $(#[$impl_attr])*
@@ -495,11 +595,11 @@ macro_rules! buffer_linear {
                 Some(Self::_new(storage, Self::_usize_to_idx(src.len())))
             }
 
-            /* common query methods */
+            /* size & capacity */
 
-            $crate::buffer_linear!(%common_owned $name, $I, $P);
+            $crate::buffer_linear!(%common_static $name, $I, $P);
 
-            /* deconstruct */
+            /* logical range control */
 
             /// Sets the logical length to zero.
             ///
@@ -593,7 +693,6 @@ macro_rules! buffer_linear {
                 let index_usize = Self::_idx_to_usize(index);
                 Some($crate::Mem::replace(&mut self.storage[index_usize], T::default()))
             }
-
             /// Takes the value at `index`, replacing it with `T::INIT`.
             pub const fn take_init(&mut self, index: $I) -> Option<T>
             where T: $crate::ConstInitCore {
@@ -601,14 +700,12 @@ macro_rules! buffer_linear {
                 let index_usize = Self::_idx_to_usize(index);
                 Some($crate::Mem::replace(&mut self.storage[index_usize], T::INIT))
             }
-
             /// Takes the value at `index`, replacing it with `other`.
             pub fn take_with(&mut self, index: $I, other: T) -> Option<T> {
                 if Self::_idx_ge(index, self.len()) { return None; }
                 let index_usize = Self::_idx_to_usize(index);
                 Some($crate::Mem::replace(&mut self.storage[index_usize], other))
             }
-
             /// Takes the value at `index`, replacing it with a copy of `other`.
             pub const fn take_with_copy(&mut self, index: $I, other: T) -> Option<T>
             where T: Copy {
@@ -619,14 +716,18 @@ macro_rules! buffer_linear {
 
             /* views */
 
-            /// Returns the initialized prefix as a slice.
+            /// Returns the active logical range as a slice.
             pub const fn as_slice(&self) -> &[T] {
                 let len = self._len_usize(); $crate::Slice::range_to(&self.storage, len)
             }
-            /// Returns the initialized prefix as an exclusive slice.
+            /// Returns the active logical range as an exclusive slice.
             pub fn as_mut_slice(&mut self) -> &mut [T] {
                 let len = self._len_usize(); $crate::Slice::range_to_mut(&mut self.storage, len)
             }
+
+            /* iteration & visitation */
+
+            $crate::buffer_linear!(%common_iter_visit $name, $I, $P);
         }
     };
     (%impl_uninit $(#[$impl_attr:meta])* $name:ident, $I:ty, $P:ty) => {
@@ -717,11 +818,11 @@ macro_rules! buffer_linear {
                 Some(Self::_new(storage, Self::_usize_to_idx(src.len())))
             }
 
-            /* common query methods */
+            /* size & capacity */
 
-            $crate::buffer_linear!(%common_owned $name, $I, $P);
+            $crate::buffer_linear!(%common_static $name, $I, $P);
 
-            /* deconstruct */
+            /* logical range control */
 
             /// Drops all initialized elements and resets the buffer.
             pub fn clear(&mut self) {
@@ -745,12 +846,19 @@ macro_rules! buffer_linear {
             /// Sets the length to `min(new_len, len)`, dropping the truncated elements.
             ///
             /// If `new_len >= len`, this is a no-op.
+            // WAIT: [const_drop_in_place](https://github.com/rust-lang/rust/issues/109342)
             pub fn truncate(&mut self, new_len: $I) {
                 while self.len() > new_len {
                     self.len = self._len_dec();
                     // SAFETY: see above
                     unsafe { self.storage[self._len_usize()].assume_init_drop(); }
                 }
+            }
+            /// Primitive-index variant of [`truncate`][Self::truncate],
+            #[inline(always)]
+            pub fn truncate_prim(&mut self, new_len: $P) -> Result<(), $crate::InvalidValue> {
+                self.truncate($crate::unwrap![ok? Self::_prim_to_idx(new_len)]);
+                Ok(())
             }
 
             /* push */
@@ -812,21 +920,10 @@ macro_rules! buffer_linear {
 
             /* take */
 
-            /// Takes the value at `index`, replacing it with `value`.
-            pub fn take_with(&mut self, index: $I, value: T) -> Option<T> {
-                if Self::_idx_ge(index, self.len()) { return None; }
-                let index_usize = Self::_idx_to_usize(index);
-                // SAFETY: `index < self.len`, so the slot is initialized per invariant.
-                let old = unsafe { self.storage[index_usize].assume_init_read() };
-                self.storage[index_usize].write(value);
-                Some(old)
-            }
-
             /// Takes the value at `index`, replacing it with `T::default()`.
             pub fn take_default(&mut self, index: $I) -> Option<T> where T: Default {
                 self.take_with(index, T::default())
             }
-
             /// Takes the value at `index`, replacing it with `T::INIT`.
             pub const fn take_init(&mut self, index: $I) -> Option<T>
             where T: $crate::ConstInitCore {
@@ -837,10 +934,19 @@ macro_rules! buffer_linear {
                 self.storage[index_usize].write(T::INIT);
                 Some(old)
             }
+            /// Takes the value at `index`, replacing it with `value`.
+            pub fn take_with(&mut self, index: $I, value: T) -> Option<T> {
+                if Self::_idx_ge(index, self.len()) { return None; }
+                let index_usize = Self::_idx_to_usize(index);
+                // SAFETY: `index < self.len`, so the slot is initialized per invariant.
+                let old = unsafe { self.storage[index_usize].assume_init_read() };
+                self.storage[index_usize].write(value);
+                Some(old)
+            }
 
             /* views */
 
-            /// Returns the initialized prefix as a slice.
+            /// Returns the active logical range as a slice.
             pub const fn as_slice(&self) -> &[T] {
                 // SAFETY: `0..self.len` is initialized per invariant.
                 unsafe { ::core::slice::from_raw_parts(
@@ -848,7 +954,7 @@ macro_rules! buffer_linear {
                     self._len_usize()
                 )}
             }
-            /// Returns the initialized prefix as a mutable slice.
+            /// Returns the active logical range as a mutable slice.
             pub const fn as_mut_slice(&mut self) -> &mut [T] {
                 // SAFETY: `0..self.len` is initialized per invariant.
                 unsafe { ::core::slice::from_raw_parts_mut(
@@ -856,7 +962,7 @@ macro_rules! buffer_linear {
                 ) }
             }
 
-            $crate::buffer_linear!(%common_array_uninit_slice_mut $name, $I, $P);
+            $crate::buffer_linear!(%common_iter_visit $name, $I, $P);
         }
     };
     (%impl_option $(#[$impl_attr:meta])* $name:ident, $I:ty, $P:ty) => {
@@ -867,7 +973,7 @@ macro_rules! buffer_linear {
         /// # Invariants
         /// - Slots `0 .. len` are `Some`
         /// - Slots `len .. CAP` are logically outside the buffer
-        /// - No holes are permitted in the initialized prefix
+        /// - No holes are permitted in the active logical range.
         ///
         /// # Notes
         /// - `Option<T>` is used to control initialization and dropping, not sparsity
@@ -923,7 +1029,7 @@ macro_rules! buffer_linear {
             ///
             /// Returns `None` if:
             /// - a `None` appears before a `Some`
-            /// - any element after the prefix is `Some`
+            /// - any element after the active logical range is `Some`
             pub fn from_array_linear(array: [Option<T>; CAP]) -> Option<Self> {
                 let mut len = 0;
                 $crate::whilst! { i in 0..CAP; {
@@ -963,11 +1069,11 @@ macro_rules! buffer_linear {
                 Some(Self::_new(storage, Self::_usize_to_idx(src.len())))
             }
 
-            /* common query methods */
+            /* size & capacity */
 
-            $crate::buffer_linear!(%common_owned $name, $I, $P);
+            $crate::buffer_linear!(%common_static $name, $I, $P);
 
-            /* deconstruct */
+            /* logical range control */
 
             /// Clears the buffer, dropping all elements.
             pub fn clear(&mut self) {
@@ -1086,12 +1192,12 @@ macro_rules! buffer_linear {
 
             /* views */
 
-            /// Returns the initialized prefix as a slice.
+            /// Returns the active logical range as a slice.
             pub const fn as_slice(&self) -> &[Option<T>] {
                 $crate::Slice::range_to(&self.storage, self._len_usize())
             }
 
-            /// Returns the initialized prefix as a mutable slice.
+            /// Returns the active logical range as a mutable slice.
             pub fn as_mut_slice(&mut self) -> &mut [Option<T>] {
                 let len_usize = self._len_usize();
                 $crate::Slice::range_to_mut(&mut self.storage, len_usize)
@@ -1124,13 +1230,13 @@ macro_rules! buffer_linear {
                     .for_each(|x| f($crate::unwrap![some_guaranteed_or_ub x.as_mut()]));
             }
 
-            /// Visits the initialized prefix as a shared slice of `Some(T)`,
+            /// Visits the active logical range as a shared slice of `Some(T)`,
             /// without exposing borrow identity.
             pub fn visit_slice<F, R>(&self, f: F)
                 -> R where for<'v> F: FnOnce(&'v [Option<T>]) -> R {
                 let len = self._len_usize(); f(&self.storage[..len])
             }
-            /// Visits the initialized prefix as an exclusive slice of `Some(T)`,
+            /// Visits the active logical range as an exclusive slice of `Some(T)`,
             /// without exposing borrow identity.
             pub fn visit_mut_slice<F, R>(&mut self, f: F)
                 -> R where for<'v> F: FnOnce(&'v mut [Option<T>]) -> R {
@@ -1152,8 +1258,7 @@ macro_rules! buffer_linear {
         ///
         /// # Notes
         /// - This type does not own its elements
-        /// - Dropping or shrinking the buffer does not drop values
-        /// - Mutations affect the underlying slice
+        /// - Shrinking the buffer does not drop values, but mutating may overwrite.
         impl<'a, T> $name<'a, T, &'a mut [T]> {
             /* construct */
 
@@ -1164,22 +1269,12 @@ macro_rules! buffer_linear {
                 if slice.len() > Self::_IDX_MAX_USIZE { return None }
                 Some(Self::_new(slice, Self::_idx_zero()))
             }
-
             /// Creates an empty buffer, truncating the backing slice
             /// if it exceeds what the index type can represent.
             pub const fn new_truncated(slice: &'a mut [T]) -> Self {
                 let cap = $crate::Cmp(slice.len()).min(Self::_IDX_MAX_USIZE);
                 let slice = $crate::Slice::range_to_mut(slice, cap);
                 Self::_new(slice, Self::_idx_zero())
-            }
-
-            /// Creates a buffer over an exclusive slice with an explicit logical length.
-            ///
-            /// Returns `None` if `len > slice.len()` or cannot be represented by the index type.
-            pub const fn from_slice_with(slice: &'a mut [T], len: $I) -> Option<Self> {
-                if slice.len() > Self::_IDX_MAX_USIZE { return None }
-                if $crate::MaybeNiche(len).to_usize_saturating() > slice.len() { return None }
-                Some(Self::_new(slice, $crate::MaybeNiche(len)))
             }
 
             /// Creates a buffer over an exclusive slice.
@@ -1189,8 +1284,15 @@ macro_rules! buffer_linear {
                 if slice.len() > Self::_IDX_MAX_USIZE { return None }
                 Some(Self::_new(slice, Self::_usize_to_idx_sat(slice.len())))
             }
-
-            /// Creates a buffer over an exclusive slice, truncating the visible prefix
+            /// Creates a buffer over an exclusive slice with an explicit logical length.
+            ///
+            /// Returns `None` if `len > slice.len()` or cannot be represented by the index type.
+            pub const fn from_slice_with(slice: &'a mut [T], len: $I) -> Option<Self> {
+                if slice.len() > Self::_IDX_MAX_USIZE { return None }
+                if $crate::MaybeNiche(len).to_usize_saturating() > slice.len() { return None }
+                Some(Self::_new(slice, $crate::MaybeNiche(len)))
+            }
+            /// Creates a buffer over an exclusive slice, truncating the visible active range
             /// if the slice length exceeds what the index type can represent.
             pub const fn from_slice_truncated(slice: &'a mut [T]) -> Self {
                 let len_usize = $crate::Cmp(slice.len()).min(Self::_IDX_MAX_USIZE);
@@ -1198,11 +1300,11 @@ macro_rules! buffer_linear {
                 Self::_new(slice, Self::_usize_to_idx_sat(len_usize))
             }
 
-            /* common query methods */
+            /* capacity */
 
             $crate::buffer_linear!(%common_view $name, $I, $P);
 
-            /* deconstruct */
+            /* logical range control */
 
             /// Sets the logical length to zero.
             ///
@@ -1214,6 +1316,11 @@ macro_rules! buffer_linear {
             /// If `new_len >= len`, this is a no-op.
             pub const fn truncate(&mut self, new_len: $I) {
                 if Self::_idx_le(new_len, self.len()) { self._set_len(new_len); }
+            }
+            /// Primitive-index variant of [`truncate`][Self::truncate],
+            pub const fn truncate_prim(&mut self, new_len: $P) -> Result<(), $crate::InvalidValue> {
+                self.truncate($crate::unwrap![ok? Self::_prim_to_idx(new_len)]);
+                Ok(())
             }
 
             /* push */
@@ -1281,17 +1388,17 @@ macro_rules! buffer_linear {
 
             /* views */
 
-            /// Returns the initialized prefix as a slice.
+            /// Returns the active logical range as a slice.
             pub const fn as_slice(&self) -> &[T] {
                 $crate::Slice::range_to(&self.storage, self._len_usize())
             }
-            /// Returns the initialized prefix as a mutable slice.
+            /// Returns the active logical range as a mutable slice.
             pub fn as_mut_slice(&mut self) -> &mut [T] {
                 let len_usize = self._len_usize();
                 $crate::Slice::range_to_mut(&mut self.storage, len_usize)
             }
 
-            $crate::buffer_linear!(%common_array_uninit_slice_mut $name, $I, $P);
+            $crate::buffer_linear!(%common_iter_visit $name, $I, $P);
         }
     };
     (%impl_slice $(#[$impl_attr:meta])* $name:ident, $I:ty, $P:ty) => {
@@ -1306,7 +1413,7 @@ macro_rules! buffer_linear {
         /// # Notes
         /// - This type does not own its elements
         /// - No mutation or removal operations are supported
-        /// - `len` limits the visible prefix
+        /// - `len` limits the visible logical range
         impl<'a, T> $name<'a, T, &'a [T]> {
             /// Creates a buffer over a shared slice.
             ///
@@ -1325,7 +1432,7 @@ macro_rules! buffer_linear {
                 Some(Self::_new(slice, $crate::MaybeNiche(len)))
             }
 
-            /// Creates a buffer over a shared slice, truncating the visible prefix
+            /// Creates a buffer over a shared slice, truncating the visible logical range
             /// if the slice length exceeds what the index type can represent.
             pub const fn from_slice_truncated(slice: &'a [T]) -> Self {
                 let len_usize = $crate::Cmp(slice.len()).min(Self::_IDX_MAX_USIZE);
@@ -1333,7 +1440,7 @@ macro_rules! buffer_linear {
                 Self::_new(slice, Self::_usize_to_idx_sat(len_usize))
             }
 
-            /* common query methods */
+            /* capacity */
 
             $crate::buffer_linear!(%common_view $name, $I, $P);
 
@@ -1355,10 +1462,13 @@ macro_rules! buffer_linear {
 
             /* views */
 
-            /// Returns the initialized prefix as a slice.
+            /// Returns the active logical range as a slice.
             pub const fn as_slice(&self) -> &[T] {
                 $crate::Slice::range_to(&self.storage, self.len.to_usize_saturating())
             }
+
+            /* iteration */
+
             /// Iterates over the initialized elements.
             pub fn iter(&self) -> impl Iterator<Item = &T> {
                 let len = self._len_usize(); self.storage[..len].iter()
@@ -1370,11 +1480,135 @@ macro_rules! buffer_linear {
             pub fn visit_each<F>(&self, f: F) where for<'v> F: Fn(&'v T) {
                 let len = self._len_usize(); for item in &self.storage[..len] { f(item); }
             }
-            /// Visits the initialized prefix as a shared slice
+            /// Visits the active logical range as a shared slice
             /// without exposing borrow identity.
             pub fn visit_slice<F, R>(&self, f: F) -> R where for<'v> F: FnOnce(&'v [T]) -> R {
                 let len = self._len_usize(); f(&self.storage[..len])
             }
+        }
+    };
+
+    /* `alloc` variants (vec) */
+
+    (%impl_vec $(#[$impl_attr:meta])* $name:ident, $I:ty, $P:ty) => {
+        $(#[$impl_attr])*
+        ///
+        /// Dynamically sized array.
+        ///
+        /// # Semantics
+        /// - Logical length is identical to `Vec::len()`.
+        /// - Capacity may grow dynamically
+        /// - `push_back` never fails
+        impl<T> $name<T, ::alloc::vec::Vec<T>> {
+            /* construct */
+
+            /// Creates a buffer from an emtpy `Vec`.
+            pub const fn new() -> Self { Self::_new(::alloc::vec::Vec::new()) }
+
+            /// Creates a buffer from an emtpy `Vec` with at least the given `capacity`.
+            pub fn with_capacity(capacity: $I) -> Self {
+                Self::_new(::alloc::vec::Vec::with_capacity(Self::_idx_to_usize(capacity)))
+            }
+            /// Creates a buffer from an emtpy `Vec` with at least the given `capacity`.
+            pub fn with_capacity_prim(capacity: $P) -> Result<Self, $crate::InvalidValue> {
+                Ok(Self::with_capacity(Self::_prim_to_idx(capacity)?))
+            }
+
+            /* size & capacity */
+
+            /// Returns the number of elements currently stored in the buffer.
+            pub const fn len(&self) -> $I {
+                Self::_usize_to_idx_sat(self.storage.capacity()).repr()
+            }
+            /// Returns the number of elements currently stored in the buffer.
+            pub const fn len_prim(&self) -> $P { self.storage.len() as $P }
+            /// Returns `true` if the buffer contains no elements.
+            pub const fn is_empty(&self) -> bool { self.storage.is_empty() }
+
+            /// Returns the capacity of the underlying `Vec`.
+            pub const fn capacity(&self) -> $I {
+                Self::_usize_to_idx_sat(self.storage.capacity()).repr()
+            }
+            /// Returns the capacity of the underlying `Vec`.
+            pub const fn capacity_prim(&self) -> $P { self.storage.capacity() as $P }
+            /// Always returns `false`, since it can never be full.
+            pub const fn is_full(&self) -> bool { false }
+
+            /* logical range control */
+
+            /// Sets the logical length to zero.
+            ///
+            /// Does not drop elements.
+            pub fn clear(&mut self) { self.storage.clear() }
+
+            /// Sets the logical length to `min(new_len, len)`.
+            ///
+            /// If `new_len >= len`, this is a no-op.
+            pub fn truncate(&mut self, new_len: $I) {
+                self.storage.truncate(Self::_idx_to_usize(new_len));
+            }
+            /// Primitive-index variant of [`truncate`][Self::truncate],
+            pub fn truncate_prim(&mut self, new_len: $P) -> Result<(), $crate::InvalidValue> {
+                self.truncate($crate::unwrap![ok? Self::_prim_to_idx(new_len)]);
+                Ok(())
+            }
+
+            /* push */
+
+            /// Removes and returns a value from the back of the buffer.
+            pub fn push_back(&mut self, value: T) { self.storage.push(value); }
+
+            /* pop */
+
+            /// Removes and returns a value from the back of the buffer.
+            pub fn pop_back(&mut self) -> Option<T> { self.storage.pop() }
+
+            /* peek */
+
+            /// Returns a shared reference to the last element without removing it.
+            pub fn peek_back(&self) -> Option<&T> { self.storage.last() }
+            /// Returns an exclusive reference to the last element without removing it.
+            pub fn peek_mut_back(&mut self) -> Option<&mut T> { self.storage.last_mut() }
+
+            /* get */
+
+            /// Returns a shared reference to the element at `index`, or `None` if out of bounds.
+            pub fn get(&self, index: $I) -> Option<&T> {
+                let idx = Self::_idx_to_usize(index);
+                self.storage.get(idx)
+            }
+            /// Returns an exclusive reference to the element at `index`,
+            /// or `None` if out of bounds.
+            pub fn get_mut(&mut self, index: $I) -> Option<&mut T> {
+                let idx = Self::_idx_to_usize(index);
+                self.storage.get_mut(idx)
+            }
+
+            /* take */
+
+            /// Takes the value at `index`, replacing it with `value`.
+            pub fn take_with(&mut self, index: $I, other: T) -> Option<T> {
+                let idx = Self::_idx_to_usize(index);
+                if idx >= self.storage.len() { return None; }
+                Some($crate::Mem::replace(&mut self.storage[idx], other))
+            }
+            /// Takes the value at `index`, replacing it with `T::default()`.
+            pub fn take_default(&mut self, index: $I) -> Option<T> where T: Default {
+                self.take_with(index, T::default())
+            }
+            /// Takes the value at `index`, replacing it with `T::INIT`.
+            pub fn take_init(&mut self, index: $I) -> Option<T> where T: $crate::ConstInitCore {
+                self.take_with(index, T::INIT)
+            }
+
+            /* views */
+
+            /// Returns the active logical range as a slice.
+            pub fn as_slice(&self) -> &[T] { &self.storage }
+            /// Returns the active logical range as an exclusive slice.
+            pub fn as_mut_slice(&mut self) -> &mut [T] { &mut self.storage }
+
+            $crate::buffer_linear!(%common_iter_visit $name, $I, $P);
         }
     };
 }