+++ src/index.rs
@@ -60,7 +60,13 @@
 // `nightly_stable_1_95`: core, alloc, std…
 #![cfg_attr(
     nightly_stable_1_95,
-    feature(assert_matches, atomic_try_update, cold_path, if_let_guard, more_conversion_trait_impls,)
+    feature(
+        assert_matches,
+        atomic_try_update,
+        cold_path,
+        if_let_guard,
+        more_conversion_trait_impls,
+    )
 )]
 #![cfg_attr(all(nightly_stable_1_95, feature = "alloc"), feature(push_mut,))]
 // #![cfg_attr(all(nightly_stable_1_95, feature = "std"), feature())]